Move g_main_loop to non-default context in cgmiClient. Add check for DBUS_SESSION_BUS_ADDRESS env var in both client and daemon. This commit should allow cgmiClient lib to run in the context of a browser (in a qt loop).

+++ hplayer/source/lib/cgmi-player.c
@@ -1066,7 +1066,7 @@ cgmi_Status cgmi_GetDuration (void *pSession, float *pDuration, cgmi_SessionType
 
       GST_INFO("Stream: %s\n", pSess->playbackURI );
       GST_INFO("Position: %lld (seconds)\n", (Duration/GST_SECOND) );
-      *pDuration = (Duration/GST_SECOND);
+      *pDuration = (float)(Duration/GST_SECOND);
 
    } while (0);
    return stat;
